segunda version

+++ src/app/shared/components/body/body.component.ts
@@ -2,7 +2,13 @@ import { IfStmt, ThisReceiver } from '@angular/compiler';
 import { Component, OnInit, ViewChild, ɵɵqueryRefresh} from '@angular/core';
 import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { count } from 'rxjs';
 import { UserServiceService } from 'src/app/user-service.service';
+import Swal from 'sweetalert2';
+
+
+
+
 
 
 @Component({
@@ -10,10 +16,13 @@ import { UserServiceService } from 'src/app/user-service.service';
   templateUrl: './body.component.html',
   styleUrls: ['./body.component.scss']
 })
+
 export class BodyComponent implements OnInit {
 
   userForm: FormGroup;
   listData: any;
+  Swal = ('sweetalert2')
+  
   
   
   constructor(private fb: FormBuilder, private userService: UserServiceService) {
@@ -31,6 +40,8 @@ export class BodyComponent implements OnInit {
 
     });
 
+    
+
   }
 
   saveOrEditUser(): void{
@@ -41,42 +52,152 @@ export class BodyComponent implements OnInit {
     let phone: HTMLInputElement =<HTMLInputElement> document.getElementById("txtPhone");
     let curp: HTMLInputElement =<HTMLInputElement> document.getElementById("txtCurp");
     let title: HTMLElement =<HTMLInputElement> document.getElementById("exampleModalLabel");
+    let d = new Date();
+    let year = d.getUTCFullYear().toString();
+    let month = d.getUTCMonth().toString().length > 1 ? d.getUTCMonth().toString() : '0' + d.getUTCMonth().toString();
+    let day = d.getDate().toString().length > 1 ? d.getDate().toString() : '0' + d.getDate().toString();
 
-    if(id.value == ''){
-      this.userService.insertUser(new Date, name.value,  address.value, phone.value, curp.value).subscribe (
+    let valid: string = this.validateControls(name.value, address.value, phone.value, curp.value);
+    
+    if(valid.length == 0){
+      if(id.value == ''){
 
-        datos => {
-          if(datos){
+        
 
+        this.userService.insertUser(year + '-' + month + '-' + day, name.value,  address.value, phone.value, curp.value).subscribe (
+
+          datas => {
+
+            if(!(datas[0] == "NO")){
+
+              let list = Object.entries(datas);
+
+
+              list.forEach((value: any, index: number) => {
+
+                const form = new FormGroup({
+
+                  txtID: new FormControl(value[1].ID),
+                  txtRegistrationDate: new FormControl(value[1].registrationDate),
+                  txtName: new FormControl(value[1].name),
+                  txtAddress: new FormControl(value[1].address),
+                  txtPhone: new FormControl(value[1].phone),
+                  txtCurp: new FormControl(value[1].curp)
+            
+                });
+    
+                this.listData.push(form.value);
+
+
+              });
+
+              Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'Contact Saved',
+                showConfirmButton: false,
+                timer: 1500
+              });
+
+            }else{
+
+              Swal.fire({
+                position: 'center',
+                icon: 'error',
+                title: 'Intern Error',
+                showConfirmButton: false,
+                timer: 1500
+              });
               
+            }
           }
-        }
 
-      )
+        );
+
+        
+        
+
+      }else{
+
+        this.userService.editUser(id.value, year + '-' + month + '-' + day, name.value,  address.value, phone.value, curp.value).subscribe (
+
+          datas => {
+
+            if(datas[0] == "OK"){
+
+
+              this.listData.forEach((value: any, index: number) => {
+                
+                if(value.txtID == id.value){
+
+                  value.txtRegistrationDate = year + '-' + month + '-' + day;
+                  value.txtName = name.value;
+                  value.txtAddress = address.value;
+                  value.txtPhone = phone.value;
+                  value.txtCurp = curp.value;
+
+                }
+
+
+              });
+
+              Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'Contact Edited',
+                showConfirmButton: false,
+                timer: 1500
+              });
+
+
+            }else{
+
+              Swal.fire({
+                position: 'center',
+                icon: 'error',
+                title: 'Intern Error',
+                showConfirmButton: false,
+                timer: 1500
+              });
+              
+            }
+          }
+
+        );
+
+        
+
+      }
+
+      
+      document.getElementById('newUserModal').click(); 
+      //window.location.reload();
+  
     }else{
 
-      this.userService.editUser(id.value, new Date(), name.value,  address.value, phone.value, curp.value).subscribe (
 
-        datos => {
-          if(datos){
-
-              
-          }
-        }
-
-      )
+      Swal.fire({
+        position: 'center',
+        icon: 'warning',
+        title: valid,
+        showConfirmButton: false,
+        timer: 1500
+      })
 
     }
 
-    document.getElementById('newUserModal').click(); 
-    window.location.reload();
-    //this.getAllUsers();
+    
 
   }
 
   ngOnInit(): void {
 
     this.getAllUsers();
+    this.initEvents();
+
+  }
+
+  initEvents(){
 
     let filter: HTMLInputElement =<HTMLInputElement> document.getElementById("txtSearch");
     filter.addEventListener('keyup', function(){
@@ -122,6 +243,8 @@ export class BodyComponent implements OnInit {
 
     }, false);
 
+    
+
   }
 
   getAllUsers(){
@@ -165,8 +288,9 @@ export class BodyComponent implements OnInit {
 
   changeTitle(){
 
+    this.resetControl();
     let title: HTMLElement =<HTMLInputElement> document.getElementById("exampleModalLabel");
-    title.innerText="New User";
+    title.innerText="New Contact";
   }
 
   geteditUser(element: any){
@@ -189,7 +313,7 @@ export class BodyComponent implements OnInit {
     curp.value=element.txtCurp;
 
     let title: HTMLElement =<HTMLInputElement> document.getElementById("exampleModalLabel");
-    title.innerText="Edit User";
+    title.innerText="Edit Contact";
     
     
 
@@ -197,17 +321,63 @@ export class BodyComponent implements OnInit {
 
   deleteUser(elemento: any){
 
-
-    this.userService.deleteUser(elemento.txtID).subscribe (
-
-      datos => {
-
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
         
+        this.userService.deleteUser(elemento.txtID).subscribe (
+
+          datas => {
+  
+            if(datas[0] == "OK"){
+
+
+              this.listData.forEach((value: any, index: number) => {
+                
+                if(value.txtID == elemento.txtID){
+
+                  this.listData.splice(index, 1);
+
+                }
+
+              });
+
+              Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'The contact been deleted',
+                showConfirmButton: false,
+                timer: 1500
+              });
+
+            }else{
+
+              Swal.fire({
+                position: 'center',
+                icon: 'error',
+                title: 'Error',
+                showConfirmButton: false,
+                timer: 1500
+              });
+
+            }
+  
+          }
+        );
+
+       
 
       }
-    );
+    });
 
-    window.location.reload();
+    //window.location.reload();
     //this.getAllUsers();
   }
 
@@ -215,6 +385,8 @@ export class BodyComponent implements OnInit {
 
 
     this.userForm.reset();
+    let csv: HTMLInputElement = <HTMLInputElement> document.getElementById("txtFileCSV");
+    csv.value="";
 
   }
 
@@ -235,20 +407,36 @@ export class BodyComponent implements OnInit {
 
     let inputFile: HTMLInputElement = <HTMLInputElement> document.getElementById("txtFileCSV");
 
+
+
     let file = inputFile.files[0];
-    let reader = new FileReader();
-    reader.onload = (e) => {
-      // Cuando el archivo se terminó de cargar
-      let lines = this.parseCSV(e.target.result);
-      
-      this.setUsersCSV(lines);
+    console.log(file);
+    if(file!=null){
 
-    };
-    // Leemos el contenido del archivo seleccionado
-    reader.readAsBinaryString(file);
+      let reader = new FileReader();
+      reader.onload = (e) => {
+        // Cuando el archivo se terminó de cargar
+        let lines = this.parseCSV(e.target.result);
+        
+        this.setUsersCSV(lines);
 
-    document.getElementById('uploadCSVModal').click(); 
+      };
+      // Leemos el contenido del archivo seleccionado
+      reader.readAsBinaryString(file);
 
+      document.getElementById('uploadCSVModal').click(); 
+    }else{
+
+      Swal.fire({
+        position: 'center',
+        icon: 'warning',
+        title: 'The CSV is required',
+        showConfirmButton: false,
+        timer: 1500
+      });
+
+
+    }
   }   
   
   setUsersCSV(output: any){
@@ -256,36 +444,123 @@ export class BodyComponent implements OnInit {
 
     let forms: FormGroup[];
     let counter: number = 0;
+    let d = new Date();
+    let year = d.getUTCFullYear().toString();
+    let month = d.getUTCMonth().toString().length > 1 ? d.getUTCMonth().toString() : '0' + d.getUTCMonth().toString();
+    let day = d.getDate().toString().length > 1 ? d.getDate().toString() : '0' + d.getDate().toString();
 
-    output.forEach((value: any, index:number) => {
     
-      if(counter > 0){
 
-        
-        this.userService.insertUser(new Date(), value[1], value[2], value[3], value[4]).subscribe(
+    
+      output.forEach((value: any, index:number) => {
 
-          datos => {
-            if(datos){
-  
-                
+        console.log(counter);
+        if(counter > 0){
+
+            this.userService.insertUser(year + '-' + month + '-' + day, value[0], value[1], value[2], value[3]).subscribe(
+
+             datas => {
+
+            if(!(datas[0] == "NO")){
+
+              let list = Object.entries(datas);
+
+
+              list.forEach((value2: any, index: number) => {
+
+                const form = new FormGroup({
+
+                  txtID: new FormControl(value2[1].ID),
+                  txtRegistrationDate: new FormControl(value2[1].registrationDate),
+                  txtName: new FormControl(value2[1].name),
+                  txtAddress: new FormControl(value2[1].address),
+                  txtPhone: new FormControl(value2[1].phone),
+                  txtCurp: new FormControl(value2[1].curp)
+            
+                });
+    
+                this.listData.push(form.value);
+
+
+              });
+
+              
+
+            }else{
+
+              
             }
           }
 
-        );
-        
+            );
+          }else{
 
-      }
-      
-      counter+=1;
-    });
+            counter+=1;
+
+          }
+          
+        
+      });
+
+      Swal.fire({
+        position: 'center',
+        icon: 'success',
+        title: 'The CSV has been upload',
+        showConfirmButton: false,
+        timer: 1500
+      });
+
+     
     
-    window.location.reload();
-    //this.getAllUsers();
+    
     
 
 
   }
 
+  validateControls(name: string, address: string, phone: string, curp: string): string{
+
+    let isValid: string = "";
+    let re = /^([A-Z][AEIOUX][A-Z]{2}\d{2}(?:0[1-9]|1[0-2])(?:0[1-9]|[12]\d|3[01])[HM](?:AS|B[CS]|C[CLMSH]|D[FG]|G[TR]|HG|JC|M[CNS]|N[ETL]|OC|PL|Q[TR]|S[PLR]|T[CSL]|VZ|YN|ZS)[B-DF-HJ-NP-TV-Z]{3}[A-Z\d])(\d)$/
+      
+    if(name.length == 0){
+
+      isValid+="The name is required \n";
+
+    }
+
+    if(address.length == 0){
+
+      isValid+="The address is required \n";
+
+    }
+
+    if(phone.toString().length == 0){
+
+      isValid += "The phone number is required \n"
+
+    }else
+    if(!(phone.toString().length == 10)){
+
+      isValid += "The phone number must be ten digits \n"
+
+    }
+
+    if(curp.toString().length == 0){
+
+      isValid += "The CURP is required \n"
+
+    }else
+    if(!curp.match(re)){
+
+      isValid += "CURP format is incorrect";
+
+    }
+
+    return isValid; 
+
+  }
+  
 }
